merge(5.1-merge-homepage-forms): Single form-logic hook behind two homepage layouts
